requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asyncmy==0.2.9
bcrypt==4.1.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
fastapi==0.109.0
filelock==3.13.1
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
identify==2.5.33
idna==3.6
libcst==1.1.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
platformdirs==4.1.0
pre-commit==3.6.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.7.0
ruff==0.1.13
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
strawberry-graphql==0.217.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
